+++ src/chromeDebug.ts
@@ -2,12 +2,12 @@
  * Copyright (C) Microsoft Corporation. All rights reserved.
  *--------------------------------------------------------*/
 
-import {ChromeDebugSession, logger, UrlPathTransformer, BaseSourceMapTransformer} from 'vscode-chrome-debug-core';
+import { ChromeDebugSession, logger, UrlPathTransformer, BaseSourceMapTransformer, telemetry } from 'vscode-chrome-debug-core';
 import * as path from 'path';
 import * as os from 'os';
-import {targetFilter} from './utils';
+import { targetFilter } from './utils';
 
-import {ChromeDebugAdapter} from './chromeDebugAdapter';
+import { ChromeDebugAdapter } from './chromeDebugAdapter';
 
 const EXTENSION_NAME = 'debugger-for-chrome';
 
@@ -31,4 +31,8 @@ ChromeDebugSession.run(ChromeDebugSession.getSession(
     }));
 
 /* tslint:disable:no-var-requires */
-logger.log(EXTENSION_NAME + ': ' + require('../../package.json').version);
+const debugAdapterVersion = require('../../package.json').version;
+logger.log(EXTENSION_NAME + ': ' + debugAdapterVersion);
+
+// __GDPR__COMMON__ "Versions.DebugAdapter" : { "classification": "SystemMetaData", "purpose": "FeatureInsight" }
+telemetry.telemetry.addCustomGlobalProperty({'Versions.DebugAdapter': debugAdapterVersion});