+++ test/int/breakOnLoad.test.ts
@@ -156,6 +156,29 @@ function runCommonTests(breakOnLoadStrategy: string) {
 
             await dc.hitBreakpointUnverified({ url, webRoot: testProjectRoot }, { path: scriptPath, line: bpLine, column: bpCol });
         });
+
+        test('Hits breakpoints on the first line of two scripts', async () => {
+            const testProjectRoot = path.join(DATA_ROOT, 'breakOnLoad_javaScript');
+            const scriptPath = path.join(testProjectRoot, 'src/script.js');
+            const script2Path = path.join(testProjectRoot, 'src/script2.js');
+
+            server = createServer({ root: testProjectRoot });
+            server.listen(7890);
+
+            const url = 'http://localhost:7890/index.html';
+
+            const bpLine = 1;
+            const bpCol = 1;
+
+            await dc.hitBreakpointUnverified({ url, webRoot: testProjectRoot }, { path: scriptPath, line: bpLine, column: bpCol });
+            await dc.setBreakpointsRequest({
+                lines: [bpLine],
+                breakpoints: [{ line: bpLine, column: bpCol }],
+                source: { path: script2Path }
+            });
+            await dc.continueRequest();
+            await dc.assertStoppedLocation('breakpoint', { path: script2Path, line: bpLine, column: bpCol });
+        });
     });
 }
 
@@ -163,17 +186,17 @@ suite('BreakOnLoad', () => {
     const DATA_ROOT = testSetup.DATA_ROOT;
 
     suite('Regex Common Tests', () => {
-        runCommonTests("regex");
+        runCommonTests('regex');
     });
 
     suite('Instrument Common Tests', () => {
-        runCommonTests("instrument");
+        runCommonTests('instrument');
     });
 
     suite('Instrument Webpack Project', () => {
         let dc: ts.ExtendedDebugClient;
         setup(() => {
-            return testSetup.setup(undefined, { breakOnLoadStrategy: "instrument" })
+            return testSetup.setup(undefined, { breakOnLoadStrategy: 'instrument' })
                 .then(_dc => dc = _dc);
         });
 
@@ -240,7 +263,7 @@ suite('BreakOnLoad', () => {
     suite('BreakOnLoad Disabled (strategy: off)', () => {
         let dc: ts.ExtendedDebugClient;
         setup(() => {
-            return testSetup.setup(undefined, { breakOnLoadStrategy: "off" })
+            return testSetup.setup(undefined, { breakOnLoadStrategy: 'off' })
                 .then(_dc => dc = _dc);
         });